Refactor book generation process: remove generateBookDataFromHtml function and integrate its logic into generateBook
